# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMS REST API",
        "description": "Human Resource Management System API Documentation",
        "contact": {
            "email": "admin@hrms.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/awards": {
            "get": {
                "tags": [
                    "Awards"
                ],
                "summary": "Get all awards",
                "operationId": "7f5be4005f577bdb90bae2647e827f6b",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "query",
                        "description": "Filter by employee",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Start date filter",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "End date filter",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of awards"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Awards"
                ],
                "summary": "Create award",
                "operationId": "99543c0756254983766cb7cca40400f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "award_name",
                                    "employee_id",
                                    "date"
                                ],
                                "properties": {
                                    "award_name": {
                                        "type": "string",
                                        "example": "Employee of the Month"
                                    },
                                    "aw_description": {
                                        "type": "string",
                                        "example": "For outstanding performance"
                                    },
                                    "awr_gift_item": {
                                        "type": "string",
                                        "example": "Certificate and Gift Voucher"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-15"
                                    },
                                    "employee_id": {
                                        "type": "string",
                                        "example": "EMP001"
                                    },
                                    "awarded_by": {
                                        "type": "string",
                                        "example": "EMP002"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Award created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/awards/{id}": {
            "get": {
                "tags": [
                    "Awards"
                ],
                "summary": "Get award details",
                "operationId": "8dca7f3f592e2150307885f04a25beec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Award details"
                    },
                    "404": {
                        "description": "Award not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Awards"
                ],
                "summary": "Update award",
                "operationId": "73405e0c226b9ec385a90380b116588c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "award_name": {
                                        "type": "string"
                                    },
                                    "aw_description": {
                                        "type": "string"
                                    },
                                    "awr_gift_item": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "employee_id": {
                                        "type": "string"
                                    },
                                    "awarded_by": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Award updated"
                    },
                    "404": {
                        "description": "Award not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Awards"
                ],
                "summary": "Delete award",
                "operationId": "901c9b07e9d7757f42a850317ae9c318",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Award deleted"
                    },
                    "404": {
                        "description": "Award not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/awards/employee/{employeeId}": {
            "get": {
                "tags": [
                    "Awards"
                ],
                "summary": "Get employee awards",
                "operationId": "fa7015d5d583e2841f28686e01f9419e",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee's awards"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}